lpc55xx_hic: move MCU-LINK support out of HIC source
