+++ records/board/mcu_link.yaml
@@ -2,3 +2,6 @@ common:
     macros:
         - MCU_LINK_SUPPORT
         - MCU_LINK_PRO_GREETING
+    sources:
+        board:
+            - source/board/mcu-link.c
